+++ aw-log/src/main/java/com/answufeng/log/AwLogListener.kt
@@ -1,0 +1,6 @@
+package com.answufeng.log
+
+fun interface AwLogListener {
+
+    fun onLog(priority: Int, tag: String?, message: String, throwable: Throwable?)
+}